Move MakeLoopedMIDINotes to a util
